core[patch]: move some to BaseLanguageModel
